@@ -2,7 +2,7 @@
   "name": "vscode-graphviz",
   "displayName": "Graphviz (dot) language support for Visual Studio Code",
   "description": "This extension provides GraphViz (dot) language support for Visual Studio Code",
-  "version": "0.0.6",
+  "version": "0.0.7",
   "publisher": "joaompinto",
   "author": "João Pinto <lamego.pinto@gmail.com>",
   "contributors": [
@@ -23,14 +23,15 @@
   "icon": "images/icon.png",
   "main": "./out/src/extension",
   "engines": {
-    "vscode": "^1.31.0"
+    "vscode": "^1.50.0"
   },
   "categories": [
     "Programming Languages",
     "Snippets"
   ],
   "activationEvents": [
-    "onLanguage:dot"
+    "onLanguage:dot",
+    "onCommand:graphviz.preview"
   ],
   "contributes": {
     "languages": [
@@ -94,17 +95,44 @@
       {
         "command": "graphviz.previewToSide",
         "title": "Open Preview to the Side",
-        "category": "Graphviz"
+        "category": "Graphviz",
+        "icon": {
+          "light": "images/light/open-preview.svg",
+          "dark": "images/dark/open-preview.svg"
+        }
       }
     ],
     "menus": {
       "editor/title": [
         {
           "command": "graphviz.previewToSide",
-          "when": "resourceLangId == dot"
+          "when": "resourceLangId == dot",
+          "group": "navigation"
         }
       ],
       "editor/context": [
+        {
+          "command": "graphviz.previewToSide",
+          "when": "resourceLangId == dot"
+        }
+      ],
+      "editor/title/context": [
+        {
+          "command": "graphviz.previewToSide",
+          "when": "resourceLangId == dot"
+        }
+      ],
+      "explorer/context": [
+        {
+          "command": "graphviz.preview",
+          "when": "resourceLangId == dot"
+        }
+      ],
+      "commandPalette": [
+        {
+          "command": "graphviz.preview",
+          "when": "resourceLangId == dot"
+        },
         {
           "command": "graphviz.previewToSide",
           "when": "resourceLangId == dot"
@@ -113,22 +141,22 @@
     }
   },
   "scripts": {
-    "vscode:prepublish": "tsc -p ./",
-    "compile": "tsc -watch -p ./",
-    "postinstall": "node ./node_modules/vscode/bin/install",
-    "test": "node ./node_modules/vscode/bin/test"
+    "vscode:prepublish": "npm run compile",
+    "watch": "tsc -watch -p ./",
+    "compile": "tsc -p ./"
   },
   "devDependencies": {
-    "@types/mocha": "^2.2.32",
-    "@types/node": "^10",
-    "@types/tmp": "^0.0.33",
-    "mocha": "^2.3.3",
-    "typescript": "^4.1.3",
+    "@types/mocha": "^8.2.1",
+    "@types/node": "^12.14.1",
+    "@types/tmp": "^0.2.0",
+    "@types/vscode": "^1.50.0",
+    "mocha": "^8.3.2",
+    "typescript": "^4.2.3"
     "vscode": "^1.1.37"
   },
   "dependencies": {
-    "@hpcc-js/wasm": "^1.0.0",
-    "opn": "^6.0.0",
-    "tmp": "^0.0.33"
+    "@hpcc-js/wasm": "^1.3.0",
+    "open": "^8.0.2",
+    "tmp": "^0.2.1"
   }
 }
